requirements.txt

Flask==2.3.3
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
mysql-connector-python==8.2.0
python-dotenv==1.0.1
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2

